Moving forward nicely with a raytrace!
Learning how to use my own API, fixing some problems as I go.
